requirements.txt
django-cleanup
os
Pillow